small coding optimize

+++ Assets/Shaders/CombinedProcedural.shader
@@ -19,8 +19,8 @@ CGINCLUDE
 #include "AutoLight.cginc"
 #include "UnityPBSLighting.cginc"
 #include "CGINC/Procedural.cginc"
-	Texture2DArray<half4> _MainTex; SamplerState sampler_MainTex;
-  Texture2DArray<half3> _LightMap; SamplerState sampler_LightMap;
+	Texture2DArray<float4> _MainTex; SamplerState sampler_MainTex;
+  Texture2DArray<float3> _LightMap; SamplerState sampler_LightMap;
 	StructuredBuffer<PropertyValue> _PropertiesBuffer;
 
 
@@ -45,8 +45,8 @@ CGINCLUDE
     }
     uv *= prop.mainScaleOffset.xy;
     uv += prop.mainScaleOffset.zw;
-    half4 spec = prop.textureIndex.z >= 0 ? _MainTex.Sample(sampler_MainTex, float3(uv, prop.textureIndex.z)) : 1;
-		half4 c = (prop.textureIndex.x >= 0 ? _MainTex.Sample(sampler_MainTex, float3(uv, prop.textureIndex.x)) : 1);
+    float4 spec = prop.textureIndex.z >= 0 ? _MainTex.Sample(sampler_MainTex, float3(uv, prop.textureIndex.z)) : 1;
+		float4 c = (prop.textureIndex.x >= 0 ? _MainTex.Sample(sampler_MainTex, float3(uv, prop.textureIndex.x)) : 1);
     c.a = usedetail < 0 ? 1 : c.a;
     if(prop.textureIndex.y >= 0){
 			o.Normal =  UnpackNormal(_MainTex.Sample(sampler_MainTex, float3(uv, prop.textureIndex.y)));
@@ -72,27 +72,27 @@ CGINCLUDE
 #define GetScreenPos(pos) ((float2(pos.x, pos.y) * 0.5) / pos.w + 0.5)
 
 
-half4 ProceduralStandardSpecular_Deferred (SurfaceOutputStandardSpecular s, float3 viewDir, out half4 outGBuffer0, out half4 outGBuffer1, out half4 outGBuffer2)
+float4 ProceduralStandardSpecular_Deferred (SurfaceOutputStandardSpecular s, float3 viewDir, out float4 outGBuffer0, out float4 outGBuffer1, out float4 outGBuffer2)
 {
     // energy conservation
     float oneMinusReflectivity;
     s.Albedo = EnergyConservationBetweenDiffuseAndSpecular (s.Albedo, s.Specular, /*out*/ oneMinusReflectivity);
     // RT0: diffuse color (rgb), occlusion (a) - sRGB rendertarget
-    outGBuffer0 = half4(s.Albedo, s.Occlusion);
+    outGBuffer0 = float4(s.Albedo, s.Occlusion);
     // RT1: spec color (rgb), smoothness (a) - sRGB rendertarget
-    outGBuffer1 = half4(s.Specular, s.Smoothness);
+    outGBuffer1 = float4(s.Specular, s.Smoothness);
     // RT2: normal (rgb), --unused, very low precision-- (a)
-    outGBuffer2 = half4(s.Normal * 0.5f + 0.5f, 0);
-    half4 emission = half4(s.Emission, 1);
+    outGBuffer2 = float4(s.Normal * 0.5f + 0.5f, 0);
+    float4 emission = float4(s.Emission, 1);
     return emission;
 }
 
 float4x4 _LastVp;
 float4x4 _NonJitterVP;
-inline half2 CalculateMotionVector(float4x4 lastvp, float3 worldPos, half2 screenUV)
+inline float2 CalculateMotionVector(float4x4 lastvp, float3 worldPos, float2 screenUV)
 {
-	half4 lastScreenPos = mul(lastvp, half4(worldPos, 1));
-	half2 lastScreenUV = GetScreenPos(lastScreenPos);
+	float4 lastScreenPos = mul(lastvp, float4(worldPos, 1));
+	float2 lastScreenUV = GetScreenPos(lastScreenPos);
 	return screenUV - lastScreenUV;
 }
 
@@ -145,26 +145,26 @@ float3 _SceneOffset;
 
 // fragment shader
 void frag_surf (v2f_surf IN,
-    out half4 outGBuffer0 : SV_Target0,
-    out half4 outGBuffer1 : SV_Target1,
-    out half4 outGBuffer2 : SV_Target2,
-    out half4 outEmission : SV_Target3,
-	out half2 outMotionVector : SV_Target4,
-  out half depth : SV_TARGET5
+    out float4 outGBuffer0 : SV_Target0,
+    out float4 outGBuffer1 : SV_Target1,
+    out float4 outGBuffer2 : SV_Target2,
+    out float4 outEmission : SV_Target3,
+	out float2 outMotionVector : SV_Target4,
+  out float depth : SV_TARGET5
 ) {
   depth = IN.pos.z;
   // prepare and unpack data
   float3 worldPos = float3(IN.worldTangent.w, IN.worldBinormal.w, IN.worldNormal.w);
   float3 worldViewDir = normalize(IN.worldViewDir);
   SurfaceOutputStandardSpecular o;
-  half3x3 wdMatrix= half3x3(normalize(IN.worldTangent.xyz), normalize(IN.worldBinormal.xyz), normalize(IN.worldNormal.xyz));
+  float3x3 wdMatrix= float3x3(normalize(IN.worldTangent.xyz), normalize(IN.worldBinormal.xyz), normalize(IN.worldNormal.xyz));
   // call surface function
   surf (IN.pack0.xy, IN.pack0.zw, IN.lightmapIndex, IN.objectIndex, o);
   o.Normal = normalize(mul(o.Normal, wdMatrix));
   outEmission = ProceduralStandardSpecular_Deferred (o, worldViewDir, outGBuffer0, outGBuffer1, outGBuffer2); //GI neccessary here!
   //Calculate Motion Vector
-  half4 screenPos = mul(_NonJitterVP, float4(worldPos, 1));
-  half2 screenUV = GetScreenPos(screenPos);
+  float4 screenPos = mul(_NonJitterVP, float4(worldPos, 1));
+  float2 screenUV = GetScreenPos(screenPos);
   outMotionVector = CalculateMotionVector(_LastVp, worldPos - _SceneOffset, screenUV);
 }
 
@@ -174,7 +174,7 @@ float3 frag_gi (v2f_surf IN) : SV_TARGET{
   float3 worldPos = float3(IN.worldTangent.w, IN.worldBinormal.w, IN.worldNormal.w);
   float3 worldViewDir = normalize(IN.worldViewDir);
   SurfaceOutputStandardSpecular o;
-  half3x3 wdMatrix= half3x3(normalize(IN.worldTangent.xyz), normalize(IN.worldBinormal.xyz), normalize(IN.worldNormal.xyz));
+  float3x3 wdMatrix= float3x3(normalize(IN.worldTangent.xyz), normalize(IN.worldBinormal.xyz), normalize(IN.worldNormal.xyz));
   // call surface function
   surf (IN.pack0.xy, IN.pack0.zw, IN.lightmapIndex, IN.objectIndex, o);
   return o.Emission;
