+++ Assets/Shaders/ProceduralInstance.shader
@@ -78,29 +78,29 @@ CGINCLUDE
 #define GetScreenPos(pos) ((float2(pos.x, pos.y) * 0.5) / pos.w + 0.5)
 
 
-half4 ProceduralStandardSpecular_Deferred (SurfaceOutputStandardSpecular s, float3 viewDir, out half4 outGBuffer0, out half4 outGBuffer1, out half4 outGBuffer2)
+float4 ProceduralStandardSpecular_Deferred (SurfaceOutputStandardSpecular s, float3 viewDir, out float4 outGBuffer0, out float4 outGBuffer1, out float4 outGBuffer2)
 {
     // energy conservation
     float oneMinusReflectivity;
     s.Albedo = EnergyConservationBetweenDiffuseAndSpecular (s.Albedo, s.Specular, /*out*/ oneMinusReflectivity);
     // RT0: diffuse color (rgb), occlusion (a) - sRGB rendertarget
-    outGBuffer0 = half4(s.Albedo, s.Occlusion);
+    outGBuffer0 = float4(s.Albedo, s.Occlusion);
 
     // RT1: spec color (rgb), smoothness (a) - sRGB rendertarget
-    outGBuffer1 = half4(s.Specular, s.Smoothness);
+    outGBuffer1 = float4(s.Specular, s.Smoothness);
 
     // RT2: normal (rgb), --unused, very low precision-- (a)
-    outGBuffer2 = half4(s.Normal * 0.5f + 0.5f, 1);
-    half4 emission = half4(s.Emission, 1);
+    outGBuffer2 = float4(s.Normal * 0.5f + 0.5f, 1);
+    float4 emission = float4(s.Emission, 1);
     return emission;
 }
 float4x4 _LastVp;
 float4x4 _NonJitterVP;
 float3 _SceneOffset;
-inline half2 CalculateMotionVector(float4x4 lastvp, half3 worldPos, half2 screenUV)
+inline float2 CalculateMotionVector(float4x4 lastvp, float3 worldPos, float2 screenUV)
 {
-	half4 lastScreenPos = mul(lastvp, half4(worldPos, 1));
-	half2 lastScreenUV = GetScreenPos(lastScreenPos);
+	float4 lastScreenPos = mul(lastvp, float4(worldPos, 1));
+	float2 lastScreenUV = GetScreenPos(lastScreenPos);
 	return screenUV - lastScreenUV;
 }
 
@@ -137,12 +137,12 @@ v2f_surf vert_surf (appdata v)
 
 // fragment shader
 void frag_surf (v2f_surf IN,
-    out half4 outGBuffer0 : SV_Target0,
-    out half4 outGBuffer1 : SV_Target1,
-    out half4 outGBuffer2 : SV_Target2,
-    out half4 outEmission : SV_Target3,
-	out half2 outMotionVector : SV_Target4,
-  out half depth : SV_TARGET5
+    out float4 outGBuffer0 : SV_Target0,
+    out float4 outGBuffer1 : SV_Target1,
+    out float4 outGBuffer2 : SV_Target2,
+    out float4 outEmission : SV_Target3,
+	out float2 outMotionVector : SV_Target4,
+  out float depth : SV_TARGET5
 ) {
 	depth = IN.pos.z;
   // prepare and unpack data
@@ -158,8 +158,8 @@ void frag_surf (v2f_surf IN,
   o.Normal = normalize(mul(o.Normal, wdMatrix));
   float3 n = o.Normal;
   outEmission = ProceduralStandardSpecular_Deferred (o, worldViewDir, outGBuffer0, outGBuffer1, outGBuffer2); //GI neccessary here!
-  half4 screenPos = mul(_NonJitterVP, float4(worldPos, 1));
-  half2 screenUV = GetScreenPos(screenPos);
+  float4 screenPos = mul(_NonJitterVP, float4(worldPos, 1));
+  float2 screenUV = GetScreenPos(screenPos);
   outMotionVector = CalculateMotionVector(_LastVp, worldPos - _SceneOffset, screenUV);
 }
 
@@ -252,7 +252,7 @@ ENDCG
                 return o;
             }
 
-            half frag (v2f i) : SV_Target
+            float frag (v2f i) : SV_Target
             {
                return distance(i.worldPos, _LightPos.xyz) / _LightPos.w;
             }
